Added basic game loop and primitive, inoptimal entity
There is currently a game loop which samples inputs at around 60 samples a second. Rendering is done as often as possible, on my PC it renders at roughly 120 fps, likely because of VSync.

The entity class implemented is very inoptimal and early. Currently it creates multiple copies of what would be the same shaders, textures, vertices, etc., and the only thing that can be done with it is set its position.

The next step is to move the responsibilities of loading assets to another class, such as a level instance, and add interactivity to the program.